move header app button kind interfaces into extensions registry
